JPA mapping

+++ src/main/java/com/tarasov/market/repository/OfferingRepository.java
@@ -5,8 +5,7 @@ import com.tarasov.market.model.Offering;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 
 @Repository
-public interface OfferingRepository extends CrudRepository<Offering, BigInteger> {
+public interface OfferingRepository extends CrudRepository<Offering, Long> {
 }
