+++ src/test/java/com/tarasov/market/OfferingRepositoryTest.java
@@ -5,6 +5,8 @@ import com.tarasov.market.repository.OfferingRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class OfferingRepositoryTest extends MarketAppApplicationTest {
@@ -15,7 +17,7 @@ public class OfferingRepositoryTest extends MarketAppApplicationTest {
     @Test
     public void addOfferingTest() {
         Offering offering = offeringRepository.save(
-                new Offering("New product", "Description", "product.img", 50000L)
+                new Offering("New product", "Description", "product.img", BigDecimal.valueOf(50000))
         );
 
         assertThat(offering)
@@ -27,7 +29,7 @@ public class OfferingRepositoryTest extends MarketAppApplicationTest {
     @Test
     public void deleteCreatedOfferingTest() {
         Offering offering = offeringRepository.save(
-                new Offering("New product", "Description", "product.img", 50000L)
+                new Offering("New product", "Description", "product.img", BigDecimal.valueOf(50000))
         );
         offeringRepository.delete(offering);
         assertThat(offeringRepository.existsById(offering.getId())).isFalse();
